+++ src/lib/MsWksGraph.cxx
@@ -2101,7 +2101,7 @@ void MsWksGraph::checkTextBoxLinks(MsWksGraphInternal::RBZone &rbZone)
       continue;
     static_cast<MsWksGraphInternal::TextBoxv4 &>(*zone).m_frame = fName;
     if (textIds.find(zone->m_ids[0]) != textIds.end()) {
-      MWAW_DEBUG_MSG(("MsWksGraph::checkTextBoxLinks: id %lX already exists\n", zone->m_ids[0]));
+      MWAW_DEBUG_MSG(("MsWksGraph::checkTextBoxLinks: id %lX already exists\n", (long unsigned int) zone->m_ids[0]));
       ok = false;
       break;
     }
@@ -2116,7 +2116,7 @@ void MsWksGraph::checkTextBoxLinks(MsWksGraphInternal::RBZone &rbZone)
        link!=nextLinks.end(); ++link) {
     if (prevLinks.find(link->second)==prevLinks.end() ||
         prevLinks.find(link->second)->second!=link->first) {
-      MWAW_DEBUG_MSG(("MsWksGraph::checkTextBoxLinks: can not find prevLinks: %lX<->%lX already exists\n", link->first, link->second));
+      MWAW_DEBUG_MSG(("MsWksGraph::checkTextBoxLinks: can not find prevLinks: %lX<->%lX already exists\n", (long unsigned int) link->first, (long unsigned int) link->second));
       ok = false;
       break;
     }
@@ -2128,7 +2128,7 @@ void MsWksGraph::checkTextBoxLinks(MsWksGraphInternal::RBZone &rbZone)
         break;
       actText = nextLinks.find(actText)->second;
       if (w++ > numLinks) {
-        MWAW_DEBUG_MSG(("MsWksGraph::checkTextBoxLinks:find a loop for id %lX\n", link->first));
+        MWAW_DEBUG_MSG(("MsWksGraph::checkTextBoxLinks:find a loop for id %lX\n", (long unsigned int) link->first));
         ok = false;
         break;
       }
