+++ src/lib/WriteNowText.cxx
@@ -908,7 +908,7 @@ shared_ptr<WriteNowTextInternal::ContentZones> WriteNowText::parseContent(WriteN
         }
       }
       if ((type == 0xb || type == 0xd) && numChar != 1) {
-        MWAW_DEBUG_MSG(("WriteNowText::parseContent: find odd size for type %x entry\n", type));
+        MWAW_DEBUG_MSG(("WriteNowText::parseContent: find odd size for type %x entry\n", (unsigned int) type));
         ascFile.addPos(pos);
         ascFile.addNote("TextData:##");
 
@@ -2005,7 +2005,7 @@ bool WriteNowText::send(std::vector<WriteNowTextInternal::ContentZone> &listZone
       break;
     }
     default:
-      MWAW_DEBUG_MSG(("WriteNowText::send: find keyword %x\n",zone.m_type));
+      MWAW_DEBUG_MSG(("WriteNowText::send: find keyword %x\n",(unsigned int) zone.m_type));
       break;
     }
